requirements.txt
Django==4.2.2
djangorestframework==3.14.0
firebase_admin==6.2.0
model_mommy==2.0.0
django-dotenv==1.4.2
black==23.3.0